pyproject.toml
[project]
name = "proxmox-mcp"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0",
    "proxmoxer>=2.0",
    "requests",
    "pyyaml",
]
